rename options --runtest to --interactive (-i) and option --run is changed to --testname (-t)
